fix: refactor classes
